requirements.txt
streamlit==1.29.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
python-dotenv==1.0.0
openai==1.3.0
pdfplumber==0.10.2
